Moved all MPI communication functions to a separate file, and streamlined some minor things in commander4.py
