+++ src/python/tod_processing.py
@@ -134,65 +134,3 @@ def process_tod(band_comm: Comm, experiment_data: DetectorTOD,
     """
     todproc_output = tod2map(band_comm, experiment_data, compsep_output, params)
     return todproc_output
-
-
-def send_tod(senders: list[int], destinations: list[int], tod_map: DetectorMap, iter_idx: int, chain_idx: int):
-    """ MPI-send the results from a single band TOD processing to another
-        destination.
-
-    Assumes the COMM_WORLD communicator.
-
-    Input:
-        band_master (bool): Whether this is the master band process.
-        tod_map (DetectorMap): The output map from process_tod for the band
-            belonging to this process.
-        iter_idx (int): The current Gibbs iteration.
-        chain_idx (int): The current chain.
-        destination (ints): The world rank of the destination
-            process.
-    """
-    logger = logging.getLogger(__name__)
-    for i in range(len(senders)):
-        if MPI.COMM_WORLD.rank == senders[i]:  # If I am sender nr i, I send the data to destination nr i.
-            logger.info(f"Rank {senders[i]} sending TOD data to {destinations[i]}")
-            MPI.COMM_WORLD.send((tod_map, iter_idx, chain_idx), dest=destinations[i])
-
-
-def receive_tod(senders, my_compsep_rank):
-    """ MPI-receive the results from the TOD processing (used in conjunction
-        with send_tod).
-
-    Input:
-        proc_master (bool): Whether this is the master of the TOD processing
-            communicator.
-        proc_comm (MPI.Comm): The TOD processing communicator.
-        senders (list of int): The sender world rank of the compsep information.
-        num_bands (int): The number of bands to receive from (should be same as
-            length of 'senders').
-
-    Returns:
-        data (list of DetectorMaps): nbands (DetectorMap)
-        iter (int): The Gibbs iteration of the data being received.
-        chain (int): The chain of the data being received.
-    """
-
-    logger = logging.getLogger(__name__)
-    # data, iter, chain = [], [], []
-    # if proc_master:
-    #     for i in range(num_bands):
-    #         _data, _iter, _chain = MPI.COMM_WORLD.recv(source=senders[i])
-    #         data.append(_data)
-    #         iter.append(_iter)
-    #         chain.append(_chain)
-        # log.logassert(np.all([i == iter[0] for i in iter]), "Different CompSep tasks received different Gibbs iteration number from TOD loop!", logger)
-        # log.logassert(np.all([i == chain[0] for i in chain]), "Different CompSep tasks received different Gibbs chain number from TOD loop!", logger)
-        # chain = chain[0]
-        # iter = iter[0]
-    
-
-    # data = proc_comm.bcast(data, root=0)
-    # iter = proc_comm.bcast(iter, root=0)
-    # chain = proc_comm.bcast(chain, root=0)
-    logger.info(f"Rank {my_compsep_rank} receiving TOD data from {senders[my_compsep_rank]}")
-    data, iter, chain = MPI.COMM_WORLD.recv(source=senders[my_compsep_rank])
-    return data, iter, chain
